+++ manager/src/components/EmployeeCreate.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
-import { Picker, Text } from 'react-native';
 import { connect } from 'react-redux';
 import { employeeUpdate, employeeCreate } from '../actions';
-import { Card, CardSection, Input, Button } from './common';
+import { Card, CardSection, Button } from './common';
+import EmployeeForm from './EmployeeForm';
 
 class EmployeeCreate extends Component {
   onButtonPress() {
@@ -13,44 +13,10 @@ class EmployeeCreate extends Component {
 
   render() {
     console.log(this.props.employee);
-    
+
     return (
       <Card>
-        <CardSection>
-          <Input
-            label="Name"
-            placeholder="Jane Doe"
-            value={this.props.name}
-            onChangeText={text => this.props.employeeUpdate({ prop: 'name', value: text })}
-             />
-        </CardSection>
-
-        <CardSection>
-          <Input
-            label="Phone"
-            placeholder="555-555-5555"
-            value={this.props.phone}
-            onChangeText={text => this.props.employeeUpdate({ prop: 'phone', value: text })}
-            />
-        </CardSection>
-
-        <CardSection style={{ flexDirection: 'column' }}>
-          <Text style={styles.pickerLabelStyle}>Shift</Text>
-          <Picker
-            selectedValue={this.props.shift}
-            onValueChange={ day => this.props.employeeUpdate({ prop: 'shift', value: day }) }
-
-          >
-            <Picker.Item label="Lunedi" value="Lunedi" />
-            <Picker.Item label="Martedi" value="Martedi" />
-            <Picker.Item label="Mercoledi" value="Mercoledi" />
-            <Picker.Item label="Jovedi" value="Jovedi" />
-            <Picker.Item label="Venerdi" value="Venerdi" />
-            <Picker.Item label="Sabado" value="Sabado" />
-            <Picker.Item label="Domenica" value="Domenica" />
-          </Picker>
-        </CardSection>
-
+        <EmployeeForm {...this.props} />
         <CardSection>
           <Button onPress={this.onButtonPress.bind(this)}>
             Create
@@ -61,13 +27,6 @@ class EmployeeCreate extends Component {
   }
 
 } // class EmployeeCreate
-
-const styles = {
-  pickerLabelStyle: {
-    fontSize: 18,
-    paddingLeft: 20
-  }
-}
 
 const mapStateToProps = (state) => {
   const { name, phone, shift } = state.employeeForm;
